requirements_v3.txt
pytesseract==0.3.10
Pillow==9.5.0
PyMuPDF==1.23.8
python-docx==1.1.0
python-pptx==0.6.22
openpyxl==3.1.2
beautifulsoup4==4.8.2
requests==2.32.3
openai==1.3.8
python-dotenv==1.0.0
google-generativeai==0.3.2
mistralai==0.4.2
protobuf==4.25.4